backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.5
httpx==0.28.1
python-multipart==0.0.20
google-generativeai==0.8.4
sse-starlette==2.1.3